+++ src/main/java/org/loverde/geographiccoordinate/Latitude.java
@@ -8,7 +8,8 @@ package org.loverde.geographiccoordinate;
 import java.util.Locale;
 
 
-public class Latitude extends GeographicCoordinate {
+public class Latitude extends GeographicCoordinateImpl {
+
    public static enum Direction {
       NORTH,
       SOUTH
@@ -18,13 +19,13 @@ public class Latitude extends GeographicCoordinate {
 
    private Direction direction;
 
+
    public Latitude() {
-      super( GeographicCoordinate.Type.LATITUDE );
+      super( GeographicCoordinateImpl.Type.LATITUDE );
    }
 
-   public Latitude( final double latitude )
-   throws GeographicCoordinateException {
-      super( GeographicCoordinate.Type.LATITUDE,
+   public Latitude( final double latitude ) throws GeographicCoordinateException {
+      super( GeographicCoordinateImpl.Type.LATITUDE,
              (int) Math.abs(latitude),
              (int) ((Math.abs(latitude) - (int)Math.abs(latitude)) * 60.0d),
              (((Math.abs(latitude) - (int)Math.abs(latitude)) * 60.0d) % 1.0d) * 60.0d );
@@ -32,9 +33,8 @@ public class Latitude extends GeographicCoordinate {
       setDirection( latitude > 0.0d ? Direction.NORTH : Direction.SOUTH );
    }
 
-   public Latitude( final int degrees, final int minutes, final double seconds, final Direction dir )
-   throws GeographicCoordinateException {
-      super( GeographicCoordinate.Type.LATITUDE, degrees, minutes, seconds );
+   public Latitude( final int degrees, final int minutes, final double seconds, final Direction dir ) throws GeographicCoordinateException {
+      super( GeographicCoordinateImpl.Type.LATITUDE, degrees, minutes, seconds );
       setDirection( dir );
    }
 
@@ -42,9 +42,7 @@ public class Latitude extends GeographicCoordinate {
     * @param direction Must be a member of {@code Latitude.Direction}
     */
    public void setDirection( final Latitude.Direction direction ) {
-      if( direction == null ) {
-         throw new IllegalArgumentException( GeographicCoordinateException.Messages.DIRECTION_NULL );
-      }
+      if( direction == null ) throw new IllegalArgumentException( GeographicCoordinateException.Messages.DIRECTION_NULL );
 
       this.direction = direction;
    }
@@ -53,10 +51,9 @@ public class Latitude extends GeographicCoordinate {
       return direction;
    }
 
+   @Override
    public double toDouble() {
-      if( getDirection() == null ) {
-         throw new IllegalStateException( GeographicCoordinateException.Messages.DIRECTION_NULL );
-      }
+      if( getDirection() == null ) throw new IllegalStateException( GeographicCoordinateException.Messages.DIRECTION_NULL );
 
       final double decimal = getDegrees() + (getMinutes() / 60.0d) + (getSeconds() / 3600.0d);
       return getDirection() == Direction.NORTH ? decimal : -decimal;